Module_B/requirements.txt
flask
pymysql
pyjwt
cryptography
matplotlib
jupyter
nbformat
ipykernel
